Moved emails to separate file
